Changing the GUID is generally NOT a good idea
The GUID is what causes the plugin to “light up” the pre-configured
plugin dashboards in the user’s New Relic account. If the GUID is
changed, the user will get plugin metrics in their NR account, but
they’ll need to build all their own dashboards - which somewhat defeats
the whole purpose of NR plugins!